requirements.txt
Flask==3.1.1
flask_sqlalchemy==3.1.1
flask_admin==1.6.1
flask_login==0.6.3
Pillow==11.2.1
WTForms==3.0.1
PyJWT==2.10.1
flask_jwt_extended==4.7.1